Move provision source to component
